WIP: Plumb through ACCP Dependency
Also, re-enable publishToMavenLocal by disabling signing of MPL.
The publishing configurations of our gradle projects are wrong.
We only need ONE publishing configuration,
with multiple repositories.

The signing logic should be determined by the repository,
not the publication.

Also, the default version in the build.gradle.kts should be 1.x-SNAPSHOT,
or something like that.

Then, when we publish to CA or Maven,
we override that.